# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Admin authentication
  - name: news
    description: News management
  - name: activities
    description: Activities management
  - name: teachers
    description: Teachers management
  - name: students
    description: Students management
  - name: contacts
    description: Contact messages
  - name: stats
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List all news
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"

  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get news by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: News article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateNews
      tags: [news]
      summary: Update news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List all activities
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Activities list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"

  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get activity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    put:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /teachers:
    get:
      operationId: listTeachers
      tags: [teachers]
      summary: List all teachers
      responses:
        "200":
          description: Teachers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherListResponse"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Create teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"

  /teachers/{id}:
    put:
      operationId: updateTeacher
      tags: [teachers]
      summary: Update teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeacherRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete teacher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: grade
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  /students/{id}:
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contact messages
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Contact messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"

  /contacts/{id}/read:
    patch:
      operationId: markContactRead
      tags: [contacts]
      summary: Mark contact message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /contacts/{id}:
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required: [token, user]

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
      required: [id, username, createdAt]

    DashboardStats:
      type: object
      properties:
        newsCount:
          type: integer
        activitiesCount:
          type: integer
        teachersCount:
          type: integer
        studentsCount:
          type: integer
        unreadContactsCount:
          type: integer
      required: [newsCount, activitiesCount, teachersCount, studentsCount, unreadContactsCount]

    NewsArticle:
      type: object
      properties:
        id:
          type: integer
        titleEn:
          type: string
        titleKh:
          type: string
        contentEn:
          type: string
        contentKh:
          type: string
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
        isPublished:
          type: boolean
        publishedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, titleEn, titleKh, contentEn, contentKh, category, isPublished, publishedAt, createdAt, updatedAt]

    CreateNewsRequest:
      type: object
      properties:
        titleEn:
          type: string
        titleKh:
          type: string
        contentEn:
          type: string
        contentKh:
          type: string
        imageUrl:
          type: string
          nullable: true
        category:
          type: string
        isPublished:
          type: boolean
      required: [titleEn, titleKh, contentEn, contentKh, category]

    NewsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        total:
          type: integer
      required: [data, total]

    Activity:
      type: object
      properties:
        id:
          type: integer
        titleEn:
          type: string
        titleKh:
          type: string
        descriptionEn:
          type: string
        descriptionKh:
          type: string
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        eventDate:
          type: string
        likes:
          type: integer
        commentsCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, titleEn, titleKh, descriptionEn, descriptionKh, category, eventDate, likes, commentsCount, createdAt, updatedAt]

    CreateActivityRequest:
      type: object
      properties:
        titleEn:
          type: string
        titleKh:
          type: string
        descriptionEn:
          type: string
        descriptionKh:
          type: string
        category:
          type: string
        imageUrl:
          type: string
          nullable: true
        eventDate:
          type: string
        likes:
          type: integer
        commentsCount:
          type: integer
      required: [titleEn, titleKh, descriptionEn, descriptionKh, category, eventDate]

    ActivityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        total:
          type: integer
      required: [data, total]

    Teacher:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameKh:
          type: string
        subjectEn:
          type: string
        subjectKh:
          type: string
        photoUrl:
          type: string
          nullable: true
        bioEn:
          type: string
          nullable: true
        bioKh:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, nameEn, nameKh, subjectEn, subjectKh, createdAt, updatedAt]

    CreateTeacherRequest:
      type: object
      properties:
        nameEn:
          type: string
        nameKh:
          type: string
        subjectEn:
          type: string
        subjectKh:
          type: string
        photoUrl:
          type: string
          nullable: true
        bioEn:
          type: string
          nullable: true
        bioKh:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required: [nameEn, nameKh, subjectEn, subjectKh]

    TeacherListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Teacher"
        total:
          type: integer
      required: [data, total]

    Student:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        nameEn:
          type: string
        nameKh:
          type: string
        grade:
          type: string
        gender:
          type: string
        enrollmentYear:
          type: integer
        phone:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, studentId, nameEn, nameKh, grade, gender, enrollmentYear, createdAt, updatedAt]

    CreateStudentRequest:
      type: object
      properties:
        studentId:
          type: string
        nameEn:
          type: string
        nameKh:
          type: string
        grade:
          type: string
        gender:
          type: string
        enrollmentYear:
          type: integer
        phone:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
      required: [studentId, nameEn, nameKh, grade, gender, enrollmentYear]

    StudentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
      required: [data, total]

    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, fullName, message, isRead, createdAt]

    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ContactMessage"
        total:
          type: integer
      required: [data, total]