# lib/api-zod/src/generated/types/activityListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from "./activity";

export interface ActivityListResponse {
  data: Activity[];
  total: number;
}