# lib/api-zod/src/generated/types/contactListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactMessage } from "./contactMessage";

export interface ContactListResponse {
  data: ContactMessage[];
  total: number;
}