# lib/api-zod/src/generated/types/createTeacherRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeacherRequest {
  nameEn: string;
  nameKh: string;
  subjectEn: string;
  subjectKh: string;
  photoUrl?: string | null;
  bioEn?: string | null;
  bioKh?: string | null;
  phone?: string | null;
  email?: string | null;
}